Add Argument to Return Mask
This is needed since most of the time, only some dimensions have to be
filtered. One of the data dimensions (typically x) is independent and
therefore it is best left to the caller to handle what to do with the mask.

But this is left as an optional choise (False)...
...
for now.