Add smoothen()
Added a rudimentary implementation to smoothen a dataset/signal
using the Savitzky–Golay filter using the scipy.signal.savgol_filter()
function.

The aim is to abstract away the implementation (as if it isn't easy
already!) by turning the window_length into a window_fraction.

A window_fraction of 1 means all datapoints are considered and
as it tends to 0, one datapoint is considered at a time (conflicts with
polyorder).