Multiple Fixes
Rewrote tests for find_outliers() using Hypothesis.
This showed up a lot of bugs that needed to be squashed at once.

- The __handle_args() method of find_outliers() was rewritten,
inspired by its counterpart in smoothen().

- As a result, the window_fraction for find_outliers() can no longer
be None. The window_length is clipped at a minimum of 1.

- Handle edge case when data cannot be split because it is too
little.

- Handle edge case in __handle_args() method of smoothen() to
replace outliers.

It's quite a handful. Quite possible this may have broken a few things.